Moved the columns and styles codes to another file.
